ucs/dijistra func
